# docs/openapi/openapi.yaml
openapi: 3.0.1
info:
    title: Microservice API Fleet
    description: Combined OpenAPI 3.0.1 specification for all microservices.
    version: 1.0.0
servers:
    - url: http://localhost:8000
...
paths:
...
    /api/tenants/register:
        post:
            tags:
                - tenant-service
            summary: RegisterWorkspace
            operationId: RegisterWorkspace
            x-category: public
            x-service: tenant-service
            x-source:
                - tenant-service/cmd/router.go:26
                - tenant-service/internal/handler/workspace_handler.go:37
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
            responses:
                "200":
                    description: Successful response
...
components:
    securitySchemes:
        BearerAuth:
            bearerFormat: JWT
            scheme: bearer
            type: http
        InternalToken:
            in: header
            name: X-Internal-Service-Token
            type: apiKey